fix: additional int64 conversions for Phase 1
- NDArray constructors: int size -> long size
- NDArray.GetAtIndex/SetAtIndex: int index -> long index
- UnmanagedStorage.GetAtIndex/SetAtIndex: int index -> long index
- ValueCoordinatesIncrementor.Next(): int[] -> long[]
- DefaultEngine.MoveAxis: int[] -> long[]

Build still failing - cascading changes needed in:
- All incrementors (NDCoordinatesIncrementor, NDOffsetIncrementor, etc.)
- NDIterator and all cast files
- UnmanagedStorage.Cloning
- np.random.shuffle, np.random.choice

Related to #584

+++ src/NumSharp.Core/Backends/Default/Indexing/Default.NonZero.cs
@@ -66,7 +66,7 @@ namespace NumSharp.Backends
             // SIMD fast path for contiguous arrays
             if (shape.IsContiguous && ILKernelGenerator.Enabled && ILKernelGenerator.VectorBits > 0)
             {
-                var flatIndices = new List<int>(Math.Max(16, size / 4));
+                var flatIndices = new List<long>(Math.Max(16, size / 4));
                 kp.FindNonZero((T*)x.Address, size, flatIndices);
                 return kp.ConvertFlatToCoordinates(flatIndices, x.shape);
             }
